pointerevents/pointerevent_pointerMove_isPrimary_same_as_pointerDown.html

<!doctype html>
<html>
    <head>
        <title>Pointer Event: pointerMove has same isPrimary as last pointerDown with the same pointerId</title>
        <meta content="text/html; charset=UTF-8" http-equiv="Content-Type"/>
        <link rel="author" title="Microsoft" href="http://www.microsoft.com/"/>
        <meta name="assert" content="The isPrimary attribute of a pointermove event must have the same value as the isPrimary attribute of the last pointerdown event with the same pointerId attribute."/>



        <!-- http://w3c-test.org/resources/testharness.js -->
	    <script src="http://w3c-test.org/resources/testharness.js"></script>
		<!-- Additional helper script for common checks across event types -->
        <script type="text/javascript" src="pointerevent_support.js"></script>
		<script type="text/javascript">

			setup({ explicit_done: true });

			var detected_pointertypes = {};
			add_completion_callback(showPointerTypes);

			var target0;
			var test_pointerEvent;
			var pointermove_event;
			var pointerdown_event;

			function run() {
				target0 = document.getElementById("target0");
				test_pointerEvent = async_test("pointermove has same isPrimary as last pointerdown");  // set up test harness

				on_event(target0, "pointermove", function (event) {
					pointermove_event = event;
					if(pointerdown_event != null) {
						test_pointerEvent.step(function () {
							assert_equals(pointermove_event.pointerId, pointerdown_event.pointerId, "For this test, pointermove.pointerId should the same as pointerdown.pointerId.");
							assert_equals(pointermove_event.isPrimary, pointerdown_event.isPrimary, "pointermove.isPrimary should the same as pointerdown.isPrimary.");
						});
						test_pointerEvent.done(); // complete test
					}
					
					detected_pointertypes[event.pointerType] = true;
				});
				
				on_event(target0, "pointerdown", function (event) {
					pointerdown_event = event;
					detected_pointertypes[event.pointerType] = true;
				});
			}
			
			function showPointerTypes() {
				var complete_notice = document.getElementById("complete-notice");
				var pointertype_log = document.getElementById("pointertype-log");
				var pointertypes = Object.keys(detected_pointertypes); pointertype_log.innerHTML = pointertypes.length ? pointertypes.join(",") : "(none)";
				if (test_pointerEvent.status == test_pointerEvent.NOTRUN
					|| test_pointerEvent.status == test_pointerEvent.TIMEOUT) {
					complete_notice.innerHTML = "Test FAIL due to time out.  This is due to the browser not supporting pointer, or the manual steps were not run to completion.";
					complete_notice.style.background = "red";
				}
				complete_notice.style.display = "block";
			}
		</script>

		<style>
			div {
				margin: 0em;
				padding: 2em;
			}

			#target0 {
				touch-action: none;
				background: yellow;
				border: 1px solid orange;
			}

			#complete-notice {
				border: 1px solid #0a0;
				display: none;
			}

			#pointertype-log {
				font-weight: bold;
			}
		</style>
	</head>
	<body onload="run()">
		<h1>Pointer Event: pointerMove has same isPrimary as last pointerDown with the same pointerId</h1>
		<h4>Test Description: 
				The isPrimary attribute of a pointermove event must have the same value as the isPrimary attribute of the last pointerdown event with the same pointerId attribute. 
		</h4>
		<br />
		<div id="target0">
			Prese and hold a mouse button,or use touch or pen to contact this element.  Move around inside the element while maintaining contact/button down. 
			Move out of or lift off of the element to complete the test.
		</div>
		<div id="complete-notice">
			<p>Test complete:  Scroll to Summary to view Pass/Fail Results.</p>
			<p>The following pointer types were detected: <span id="pointertype-log"></span>.</p>
		</div>

		<div id="log"></div>
	</body>
</html>
